requirements.txt
pandas==2.2.3
requests==2.31.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
boto3==1.34.131
pyarrow==16.1.0
fastparquet==2024.5.0
tenacity==8.3.0
PyYAML==6.0.2
